ulong to uint32_t everywhere

+++ Lawn/System/PoolEffect.cpp
@@ -18,7 +18,7 @@ void PoolEffect::PoolEffectInitialize()
     mCausticImage = new MemoryImage(gSexyAppBase);
     mCausticImage->mWidth = CAUSTIC_IMAGE_WIDTH;
     mCausticImage->mHeight = CAUSTIC_IMAGE_HEIGHT;
-    mCausticImage->mBits = new unsigned long[CAUSTIC_IMAGE_WIDTH * CAUSTIC_IMAGE_HEIGHT + 1];
+    mCausticImage->mBits = new uint32_t[CAUSTIC_IMAGE_WIDTH * CAUSTIC_IMAGE_HEIGHT + 1];
     mCausticImage->mHasTrans = true;
     mCausticImage->mHasAlpha = true;
     memset(mCausticImage->mBits, 0xFF, CAUSTIC_IMAGE_WIDTH * CAUSTIC_IMAGE_HEIGHT * 4);
@@ -75,7 +75,7 @@ void PoolEffect::UpdateWaterEffect()
 
         for (int x = 0; x < CAUSTIC_IMAGE_WIDTH; x++)
         {
-            unsigned long* pix = &mCausticImage->mBits[idx];
+            uint32_t* pix = &mCausticImage->mBits[idx];
 
             int timeU = x << 17;
             int timePool0 = mPoolCounter << 16;
