+++ SexyAppFramework/graphics/ImageFont.cpp
@@ -48,20 +48,20 @@ ListDataElement::ListDataElement()
 
 ListDataElement::~ListDataElement()
 {
-	for (ulong i = 0; i < mElementVector.size(); i++)
+	for (uint32_t i = 0; i < mElementVector.size(); i++)
 		delete mElementVector[i];
 }
 
 ListDataElement::ListDataElement(const ListDataElement& theListDataElement)
 {
 	mIsList = true;
-	for (ulong i = 0; i < theListDataElement.mElementVector.size(); i++)
+	for (uint32_t i = 0; i < theListDataElement.mElementVector.size(); i++)
 		mElementVector.push_back(theListDataElement.mElementVector[i]->Duplicate());
 }
 
 ListDataElement& ListDataElement::operator=(const ListDataElement& theListDataElement)
 {
-	ulong i;
+	uint32_t i;
 
 	for (i = 0; i < mElementVector.size(); i++)
 		delete mElementVector[i];
@@ -86,7 +86,7 @@ CharData::CharData()
 	mWidth = 0;
 	mOrder = 0;
 
-	//for (ulong i = 0; i < 256; i++)
+	//for (uint32_t i = 0; i < 256; i++)
 	//	mKerningOffsets[i] = 0;
 }
 
@@ -131,7 +131,7 @@ FontLayer::FontLayer(const FontLayer& theFontLayer) :
 	// @Minerscale mUseAlphaCorrection wasn't initialised in the copy constructor causing UB
 	mUseAlphaCorrection(theFontLayer.mUseAlphaCorrection)
 {
-	//ulong i;
+	//uint32_t i;
 	//
 	//for (i = 0; i < 256; i++)
 	//	mCharData[i] = theFontLayer.mCharData[i];	
@@ -161,7 +161,7 @@ FontData::FontData()
 	mRefCount = 0;
 	mDefaultPointSize = 0;
 
-	//for (ulong i = 0; i < 256; i++)
+	//for (uint32_t i = 0; i < 256; i++)
 	//	mCharMap[i] = (uchar) i;
 }
 
@@ -337,7 +337,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 
 				ListDataElement* aRectList = new ListDataElement();
 
-				for (ulong aWidthNum = 0; aWidthNum < aWidthsVector.size(); aWidthNum++)
+				for (uint32_t aWidthNum = 0; aWidthNum < aWidthsVector.size(); aWidthNum++)
 				{
 					ListDataElement* aRectElement = new ListDataElement();
 					aRectList->mElementVector.push_back(aRectElement);
@@ -403,7 +403,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 			{
 				if (aFromVector.size() == aToVector.size())
 				{
-					for (ulong aMapIdx = 0; aMapIdx < aFromVector.size(); aMapIdx++)
+					for (uint32_t aMapIdx = 0; aMapIdx < aFromVector.size(); aMapIdx++)
 					{
 						if ((aFromVector[aMapIdx].length() == 1) && (aToVector[aMapIdx].length() == 1))
 						{
@@ -479,7 +479,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 			if ((DataToLayer(theParams.mElementVector[1], &aLayer)) &&
 				(DataToStringVector(theParams.mElementVector[2], &aStringVector)))
 			{
-				for (ulong i = 0; i < aStringVector.size(); i++)
+				for (uint32_t i = 0; i < aStringVector.size(); i++)
 					aLayer->mRequiredTags.push_back(StringToUpper(aStringVector[i]));
 			}
 			else
@@ -498,7 +498,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 			if ((DataToLayer(theParams.mElementVector[1], &aLayer)) &&
 				(DataToStringVector(theParams.mElementVector[2], &aStringVector)))
 			{
-				for (ulong i = 0; i < aStringVector.size(); i++)
+				for (uint32_t i = 0; i < aStringVector.size(); i++)
 					aLayer->mExcludedTags.push_back(StringToUpper(aStringVector[i]));
 			}
 			else
@@ -763,7 +763,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 			{
 				if (aCharsVector.size() == aCharWidthsVector.size())
 				{
-					for (ulong i = 0; i < aCharsVector.size(); i++)
+					for (uint32_t i = 0; i < aCharsVector.size(); i++)
 					{
 						char32_t first_char = UTF8CharToUTF32Char(aCharsVector[i]);
 						aLayer->GetCharData(first_char)->mWidth = aCharWidthsVector[i];
@@ -832,7 +832,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 						int anImageWidth = aLayer->mImage->GetWidth();
 						int anImageHeight = aLayer->mImage->GetHeight();
 
-						for (ulong i = 0; i < aCharsVector.size(); i++)
+						for (uint32_t i = 0; i < aCharsVector.size(); i++)
 						{
 							IntVector aRectElement;
 							//std::wstring aWString = UTF8StringToWString(aCharsVector[i]);
@@ -901,7 +901,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 			{
 				if (aCharsVector.size() == aRectList.mElementVector.size())
 				{
-					for (ulong i = 0; i < aCharsVector.size(); i++)
+					for (uint32_t i = 0; i < aCharsVector.size(); i++)
 					{
 						IntVector aRectElement = IntVector();
 						//std::wstring aWString = UTF8StringToWString(aCharsVector[i]);
@@ -942,7 +942,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 			{
 				if (aPairsVector.size() == anOffsetsVector.size())
 				{
-					for (ulong i = 0; i < aPairsVector.size(); i++)
+					for (uint32_t i = 0; i < aPairsVector.size(); i++)
 					{
 						std::wstring aWString = UTF8StringToWString(aPairsVector[i]);
 						if (aWString.length() == 2)
@@ -999,7 +999,7 @@ bool FontData::HandleCommand(const ListDataElement& theParams)
 			{
 				if (aCharsVector.size() == aCharOrdersVector.size())
 				{
-					for (ulong i = 0; i < aCharsVector.size(); i++)
+					for (uint32_t i = 0; i < aCharsVector.size(); i++)
 					{
 						if (aCharsVector[i].length() == 1)
 						{
@@ -1256,7 +1256,7 @@ void ImageFont::GenerateActiveFontLayers()
 
 	mActiveLayerList.clear();
 
-	ulong i;
+	uint32_t i;
 
 	mAscent = 0;
 	mAscentPadding = 0;
@@ -1386,7 +1386,7 @@ void ImageFont::GenerateActiveFontLayers()
 					if (mForceScaledImagesWhite)
 					{
 						int aCount = aMemoryImage->mWidth * aMemoryImage->mHeight;
-						ulong* aBits = aMemoryImage->GetBits();
+						uint32_t* aBits = aMemoryImage->GetBits();
 
 						for (int i = 0; i < aCount; i++) {
 							// *(aBits++) = *aBits | 0x00FFFFFF; ambiguous
@@ -1558,7 +1558,7 @@ void ImageFont::DrawStringEx(Graphics* g, int theX, int theY, const SexyString& 
 	int aCurXPos = theX;
 	int aCurPoolIdx = 0;
 
-	for (ulong aCharNum = 0; aCharNum < theString.length(); aCharNum++)
+	for (uint32_t aCharNum = 0; aCharNum < theString.length(); aCharNum++)
 	{
 		SexyChar aChar = GetMappedChar(theString[aCharNum]);//mFontData->mCharMap[(uchar) theString[aCharNum]];
